需求修改完成

See merge request wzmte-cw/cw-zzj!2

+++ src/main/java/com/zzj/dto/req/RecordDetailReqDTO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 public class RecordDetailReqDTO {
 
-    private Integer relExamId;
+    private Long relExamId;
     private String relAnswer;
     private Integer relCorrect;
     private Integer relReId;
